fix: devtool 选项无法修改
