+++ packages/vue-cli-plugin-mpx/commands/serve/mp.js
@@ -1,28 +1,27 @@
 const webpack = require('webpack')
 const { parseTarget, getCurrentTarget } = require('../../utils/index')
 const { symlinkTargetConfig } = require('../../utils/symlinkTargetConfig')
-const { resolveWebpackConfigByTarget, handleWebpackDone } = require('../../utils/webpack')
+const {
+  resolveWebpackConfigByTarget,
+  handleWebpackDone
+} = require('../../utils/webpack')
 
 const resolveMpServeWebpackConfig = (api, options, args) => {
   const customMpxEnv = args.env
   const target = parseTarget(args.target, options)
   // 小程序业务代码构建配置
-  const webpackConfigs = resolveWebpackConfigByTarget(
-    api,
-    options,
-    target,
-    (webpackConfig) => {
-      webpackConfig.devtool('source-map')
-      if (customMpxEnv) {
-        webpackConfig.plugin('mpx-webpack-plugin').tap((args) => {
-          args[0].env = customMpxEnv
-          return args
-        })
-      }
+  api.chainWebpack((config) => {
+    if (customMpxEnv) {
+      config.plugin('mpx-webpack-plugin').tap((args) => {
+        args[0].env = customMpxEnv
+        return args
+      })
     }
-  )
-  return webpackConfigs
+  })
+  return resolveWebpackConfigByTarget(api, options, target)
 }
+
+module.exports.resolveMpServeWebpackConfig = resolveMpServeWebpackConfig
 
 /** @type {import('@vue/cli-service').ServicePlugin} */
 module.exports.registerMpServeCommand = function registerMpServeCommand (
@@ -35,10 +34,12 @@ module.exports.registerMpServeCommand = function registerMpServeCommand (
     const webpackConfigs = resolveMpServeWebpackConfig(api, options, args)
     return new Promise((resolve, reject) => {
       webpack(webpackConfigs).watch({}, (err, stats) => {
-        handleWebpackDone(err, stats, target).then((...res) => {
-          symlinkTargetConfig(api, target, webpackConfigs[0])
-          resolve(...res)
-        }).catch(reject)
+        handleWebpackDone(err, stats, target)
+          .then((...res) => {
+            symlinkTargetConfig(api, target, webpackConfigs[0])
+            resolve(...res)
+          })
+          .catch(reject)
       })
     })
   })
