+++ packages/vue-cli-plugin-mpx/config/transformMpxEntry.js
@@ -1,11 +1,17 @@
-// TODO: 看是否能在 @mpxjs/webpack-plugin 内部做下处理？
-module.exports.transformMpxEntry = function transformMpxEntry (api, options = {}, webpackConfig, isWeb = false) {
+module.exports.transformMpxEntry = function transformMpxEntry (
+  api,
+  options = {},
+  webpackConfig,
+  isWeb = false
+) {
   // 通过 cli 生成的默认的入口文件
   let basePath = 'src/app.mpx'
 
   if (api.hasPlugin('mpx-cloud-func') || api.hasPlugin('mpx-plugin-mode')) {
     try {
-      const projectConfigJson = require(api.resolve('static/wx/project.config.json'))
+      const projectConfigJson = require(api.resolve(
+        'static/wx/project.config.json'
+      ))
       basePath = `src/${projectConfigJson.miniprogramRoot}/app.mpx`
     } catch (e) {}
   }
